Added relation between user and place tables

+++ server/controllers/places-controller.js
@@ -1,11 +1,14 @@
+const mongoose = require('mongoose');
 const { validationResult } = require('express-validator');
 
 const HttpError = require('../models/http-error');
 const getCoordsForAddress = require('../util/location');
 const Place = require('../models/place');
+const User = require('../models/user');
 
 const getPlaceById = async (req, res, next) => {
     const placeId = req.params.pid;
+    
     let place;
     try {
         place = await Place.findById(placeId);
@@ -15,26 +18,26 @@ const getPlaceById = async (req, res, next) => {
 
     if (!place) {
         return next(new HttpError('Could not find a place for the provided place id.', 404));
-    } 
+    }
 
-    res.json({place: place.toObject( { getters: true } ) });
+    res.json({ place: place.toObject({ getters: true }) });
 }
 
 const getPlacesByUserId = async (req, res, next) => {
     const userId = req.params.uid;
-    let places;
-
+    
+    let userWithPlaces;
     try {
-        places = await Place.find({ creator: userId });
+        userWithPlaces = await User.findById(userId).populate('places');
     } catch {
         return next(new HttpError('Fetching places failed, pleace try again', 500));
     }
 
-    if (!places || places.length === 0) {
+    if (!userWithPlaces || userWithPlaces.places.length === 0) {
         return next(new HttpError('Could not find a place for the provided user id.', 404));
-    } 
+    }
 
-    res.json({places: places.map(place => place.toObject({ getters: true }) )} );
+    res.json({ places: userWithPlaces.places.map(place => place.toObject({ getters: true })) });
 }
 
 const createPlace = async (req, res, next) => {
@@ -51,7 +54,7 @@ const createPlace = async (req, res, next) => {
     } catch (error) {
         return next(error);
     }
-    const createdPlace = new Place({
+    const place = new Place({
         title,
         description,
         address,
@@ -60,13 +63,34 @@ const createPlace = async (req, res, next) => {
         creator
     });
 
+    let user;
+
     try {
-        await createdPlace.save();
+        user = await User.findById(creator);
     } catch {
-        return next(new HttpError('Creating place failed, pleace try again.', 500));
+        return next(new HttpError('Creating place failed, pleace try again', 500));
     }
 
-    res.status(201).json({ place: createdPlace.toObject( {getters: true} )});
+    if (!user) {
+        return next(new HttpError('Could not find user for provided id.', 404));
+    }
+
+    try {
+        const session = await mongoose.startSession();
+        session.startTransaction();
+        // await place.save({session});  // does not working on single node
+        await place.save();
+        user.places.push(place);
+        // await user.save({session}); // does not working on single node
+        await user.save();
+        await session.commitTransaction();
+        await session.endSession();
+    } catch (err) {
+        console.log(err)
+        return next(new HttpError('Creating place failed, pleace try again', 500));
+    }
+
+    res.status(201).json({ place: place.toObject({ getters: true }) });
 }
 
 const updatePlace = async (req, res, next) => {
@@ -74,7 +98,7 @@ const updatePlace = async (req, res, next) => {
     if (!errors.isEmpty()) {
         return next(new HttpError('Invalid inputs passed, please check your data.', 422));
     }
-    
+
     const { title, description } = req.body;
     const placeId = req.params.pid;
 
@@ -84,7 +108,11 @@ const updatePlace = async (req, res, next) => {
     } catch {
         return next(new HttpError('Something went wrong, could not update place.', 500));
     }
-    
+
+    if (!place) {
+        return next(new HttpError('Could not find place for provided id.', 404));
+    }
+
     place.title = title;
     place.description = description;
 
@@ -94,7 +122,7 @@ const updatePlace = async (req, res, next) => {
         return next(new HttpError('Something went wrong, could not update place.', 500));
     }
 
-    res.status(200).json({ place: place.toObject( {getters: true} )});
+    res.status(200).json({ place: place.toObject({ getters: true }) });
 };
 
 const deletePlace = async (req, res, next) => {
@@ -102,18 +130,30 @@ const deletePlace = async (req, res, next) => {
 
     let place;
     try {
-        place = await Place.findById(placeId);
-    } catch {
+        place = await Place.findById(placeId).populate('creator');
+    } catch (err) {
         return next(new HttpError('Something went wrong, could not delete place.', 500));
     }
-    console.log(place)
+    
+    if (!place) {
+        return next(new HttpError('Could not find place for provided id.', 404));
+    }
+    
     try {
+        const session = await mongoose.startSession();
+        session.startTransaction();
+        // await place.remove({session}); // does not working on single node
         await place.deleteOne();
+        place.creator.places.pull(place); 
+        // await  place.creator.save({session}); // does not working on single node
+        await  place.creator.save();
+        await session.commitTransaction();
+        await session.endSession();
     } catch {
         return next(new HttpError('Something went wrong, could not delete place.', 500));
     }
 
-    res.status(200).json({ message: 'Deleted place'});
+    res.status(200).json({ message: 'Deleted place' });
 };
 
 exports.getPlaceById = getPlaceById;
